+++ test/component/inventory_view_test.dart
@@ -5,7 +5,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:spare_parts/entities/inventory_item.dart';
 import 'package:spare_parts/pages/home_page/home_page.dart';
-import 'package:spare_parts/pages/home_page/inventory_view.dart';
+import 'package:spare_parts/pages/home_page/inventory_view/inventory_view.dart';
 import 'package:spare_parts/utilities/constants.dart';
 
 import '../test_helpers.dart';
